Refactor app.js to be more lightweight, moved functionality to modules
